organized notepad, todolist,  register, login
transferred notepad elements to notepad components and restructured the rest
