+++ client/src/hooks/useToDoList.ts
@@ -364,23 +364,23 @@ export const useToDoList = () => {
     }
   };
   return {
-    afterloading,
-    setTasks,
+    setAfterLoading,
     setTasksBackup,
     setFilterType,
-    tasks,
-    tasksBackup,
-    loading,
-    filterType,
-    addTask,
-    deleteTask,
-    toggleCompleteTask,
-    saveEditedTask,
-    filterArr,
     setFilterArr,
-    setAfterLoading,
-    completedTasks,
+    setTasks,
+    afterloading,
+    tasksBackup,
+    filterType,
+    filterArr,
+    tasks,
+    loading,
+    addTask,
     afterMark,
+    deleteTask,
+    saveEditedTask,
+    completedTasks,
     taskInputDisable,
+    toggleCompleteTask,
   };
 };
