+++ client/src/pages/ToDoListPage.tsx
@@ -1,9 +1,10 @@
+import ToDoListComponent from "../components/ToDoListComponent";
 import WhiteContainer from "../components/WhiteContainer"
 import Sidebar from "../components/SideBar";
 import Avatar from "../components/Avatar";
 import Clock from "../components/Clock";
 
-import ToDoListComponent from "../components/ToDoListComponent";
+
